Move App and AppBar into components.
