+++ src/components/App/AppBar.scss
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-@import '../../common/utils';
+@import '../common/utils';
 
 .AppBar {
   background-color: var(--mdc-theme-surface);
