+++ src/components/App/AppBar.test.tsx
@@ -18,7 +18,7 @@ import { render } from '@testing-library/react';
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
 
-import { AppBar } from './index';
+import { AppBar } from './AppBar';
 
 function isTabActive(labelEl) {
   const tabEl = labelEl.closest('.mdc-tab');
