+++ src/components/App/index.scss
@@ -1,5 +1,5 @@
-@import './components/common/AppBar/variables';
-@import './App_variables.scss';
+@import './AppBar_variables';
+@import './variables';
 
 $page-gutter-size: 48px;
