+++ src/components/App/variables.scss
@@ -1,4 +1,4 @@
-@import './components/common/AppBar/variables';
+@import './AppBar_variables';
 $app-main-real-vert-padding: 24px;
 
 $app-remaining-viewport-height-for-content: calc(
