+++ src/components/Firestore/index.scss
@@ -16,7 +16,7 @@
 
 @import '../common/color';
 @import '../common/utils';
-@import '../App_variables';
+@import '../App/variables';
 
 .Firestore {
   display: flex;
